st_management/requirements.txt
django-5.1.3
django-compressor-4.5.1